requirements.txt
pandas
pytest
streamlit
google-generativeai
